requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
passporteye==2.2.1
pdf417decoder==1.0.8
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.12
